#2 moved Property Selling Point wiget to his own file
